modularizado parte 1
