+++ NEWBIE/TODO tip-calculator-app-main/index.html
@@ -90,6 +90,6 @@
         </article>
       </section>
     </main>
-    <script src="main.js"></script>
+    <script src="main.js" type="module"></script>
   </body>
 </html>
