+++ NEWBIE/TODO tip-calculator-app-main/variables.js
@@ -1,0 +1,7 @@
+export const intoBill = document.getElementById("invoice");
+export const button = document.querySelectorAll(".content-options-tips button");
+export const numberOfPeople = document.getElementById("num-people");
+export const customTip = document.getElementById("custom");
+export const resultTipAmount = document.getElementById("result-tip");
+export const resultTotalAmount = document.getElementById("result-total");
+export const resetButton = document.getElementById("reset-all");
